test: Move config's http endpoint test to config.py
WAZO-4352
